requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
textual==0.22.0
chromadb==0.3.23
python-dotenv==1.0.0
requests==2.31.0
google-genai==0.1.0